PP-461: Create PTL test directory structure as per guidelines
