+++ test/tests/functional/pbs_hook_debug_nocrash.py
@@ -36,10 +36,10 @@
 # "PBS Professional®", and "PBS Pro™" and Altair’s logos is subject to Altair's
 # trademark licensing policies.
 
-from ptl.utils.pbs_testsuite import *
+from tests.functional import *
 
 
-class TestHookDebugNoCrash(PBSTestSuite):
+class TestHookDebugNoCrash(TestFunctional):
 
     """
     This tests to make sure the following does not occur:
@@ -81,7 +81,7 @@ class TestHookDebugNoCrash(PBSTestSuite):
             msg += " to be <= %d " % self.open_files_limit_expected
             msg += "(current value=%s)." % open_files_limit
             self.skipTest(msg)
-        PBSTestSuite.setUp(self)
+        TestFunctional.setUp(self)
 
     def test_hook_debug_no_crash(self):
