+++ test/tests/functional/pbs_hookswig.py
@@ -36,33 +36,48 @@
 # "PBS Professional®", and "PBS Pro™" and Altair’s logos is subject to Altair's
 # trademark licensing policies.
 
-from ptl.utils.pbs_testsuite import *
+from tests.functional import *
 
 
-class TestPp369(PBSTestSuite):
+class TestHookSwig(TestFunctional):
 
-    def test_job_state_count(self):
+    """
+    This test suite contains hook test to verify that the internal,
+    swig-generated 'pbs_ifl.py' does not cause an exception.
+    """
+
+    def test_hook(self):
         """
-        test_job_state_count: Testing if jobs in the 'W' state
-        will cuase the state_count to go negative or incorrect.
+        Create a hook, import a hook content that test pbs.server() call.
         """
-        # Failing stage-in operation, to put job into the waiting state
-        a = {ATTR_stagein: 'inputData@' +
-             self.server.hostname + ':/noDir/nofile'}
-        j = Job(TEST_USER, a)
-        jid = self.server.submit(j)
-        self.server.expect(JOB, {'job_state': 'W'}, id=jid)
-        # Restart server
-        self.server.restart()
-        qstat = self.server.status(SERVER)
-        self.logger.info("qstat[0]['state_count']:" + qstat[0]['state_count'])
-        state_count = qstat[0]['state_count'].split()
-        all_state_count = 0
-        for s in state_count:
-            state = s.split(':')
-            # Check for negative value
-            self.assertGreaterEqual(int(state[1]), 0)
-            all_state_count = all_state_count + int(state[1])
-        self.logger.info("qstat[0]['total_jobs']:" + qstat[0]['total_jobs'])
-        self.logger.info("all_state_count = %d", all_state_count)
-        self.assertEqual(int(qstat[0]['total_jobs']), all_state_count)
+        hook_name = "testhook"
+        hook_body = """
+import pbs
+e = pbs.event()
+s = pbs.server()
+if e.job.id:
+    jid = e.job.id
+else:
+    jid = "newjob"
+pbs.logjobmsg(jid, "server is %s" % (s.name,))
+"""
+        a = {'event': 'queuejob,execjob_begin', 'enabled': 'True'}
+        self.server.create_import_hook(hook_name, a, hook_body)
+
+        j = Job(TEST_USER)
+        a = {'Resource_List.select': '1:ncpus=1', 'Resource_List.walltime': 30}
+
+        j.set_attributes(a)
+        j.set_sleep_time(10)
+
+        try:
+            jid = self.server.submit(j)
+        except PbsSubmitError:
+            pass
+        rv = self.server.log_match("Job;%s;server is %s" % (
+            "newjob", self.server.shortname), max_attempts=3)
+        self.assertTrue(rv)
+
+        rv = self.mom.log_match("Job;%s;server is %s" % (
+            jid, self.server.shortname), max_attempts=3)
+        self.assertTrue(rv)
